+++ src/module/apps/_module.mjs
@@ -1,4 +1,3 @@
 export * as CardsSheets from "./CardsSheet.mjs";
 export {default as CardSheet} from "./CardSheet.mjs";
 export {default as CardHud} from "./CardHud.mjs";
-export * as CardsDialogs from "./cards-dialogs.mjs";